// client/src/global.types.ts
export interface ItemType {
	_id?: string;
	name: string;
	start_price: string;
	current_price: string;
	time: string;
}